+++ src/Symfony/Bundle/FrameworkBundle/Test/TestContainer.php
@@ -13,12 +13,13 @@ namespace Symfony\Bundle\FrameworkBundle\Test;
 
 use Psr\Container\ContainerInterface;
 use Symfony\Component\DependencyInjection\Container;
+use Symfony\Component\DependencyInjection\Exception\InvalidArgumentException;
 use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
 use Symfony\Component\HttpKernel\KernelInterface;
 
 /**
  * A special container used in tests. This gives access to both public and
- * private services. The container will not include private services that has
+ * private services. The container will not include private services that have
  * been inlined or removed. Private services will be removed when they are not
  * used by other services.
  *
@@ -28,13 +29,11 @@ use Symfony\Component\HttpKernel\KernelInterface;
  */
 class TestContainer extends Container
 {
-    private KernelInterface $kernel;
-    private string $privateServicesLocatorId;
-
-    public function __construct(KernelInterface $kernel, string $privateServicesLocatorId)
-    {
-        $this->kernel = $kernel;
-        $this->privateServicesLocatorId = $privateServicesLocatorId;
+    public function __construct(
+        private KernelInterface $kernel,
+        private string $privateServicesLocatorId,
+        private array $renamedIds = [],
+    ) {
     }
 
     public function compile()
@@ -69,7 +68,20 @@ class TestContainer extends Container
 
     public function set(string $id, mixed $service)
     {
-        $this->getPublicContainer()->set($id, $service);
+        $container = $this->getPublicContainer();
+        $renamedId = $this->renamedIds[$id] ?? $id;
+
+        try {
+            $container->set($renamedId, $service);
+        } catch (InvalidArgumentException $e) {
+            if (!str_starts_with($e->getMessage(), "The \"$renamedId\" service is private")) {
+                throw $e;
+            }
+            if (isset($container->privates[$renamedId])) {
+                throw new InvalidArgumentException(sprintf('The "%s" service is already initialized, you cannot replace it.', $id));
+            }
+            $container->privates[$renamedId] = $service;
+        }
     }
 
     public function has(string $id): bool
@@ -104,11 +116,7 @@ class TestContainer extends Container
 
     private function getPublicContainer(): Container
     {
-        if (null === $container = $this->kernel->getContainer()) {
-            throw new \LogicException('Cannot access the container on a non-booted kernel. Did you forget to boot it?');
-        }
-
-        return $container;
+        return $this->kernel->getContainer() ?? throw new \LogicException('Cannot access the container on a non-booted kernel. Did you forget to boot it?');
     }
 
     private function getPrivateContainer(): ContainerInterface